--- src/components/Dashboard/Content/index.js
@@ -1,3 +1,0 @@
-import Content from './Content';
-
-export default Content;
