+++ src/components/organisms/DailyWorkout/Data.js
@@ -1,0 +1,11 @@
+const Data = [
+  {
+    id: '0',
+    type: 'Trainingsplan',
+    title: 'Titel des Wörkouts',
+    program: 'Titel des Programms',
+    stats: 'XXX kcal · 26 Min. · Beweglichkeit',
+  },
+];
+
+export default Data;
