+++ src/components/organisms/DailyWorkout/assets/img--dashboard-l.png
