+++ src/components/organisms/DailyWorkout/assets/img--dashboard.png
