+++ src/components/organisms/DailyWorkout/index.js
@@ -1,0 +1,3 @@
+import Content from './DailyWorkout';
+
+export default Content;
