+++ src/components/organisms/NavBar/assets/icon--browse.png
