+++ src/components/organisms/NavBar/assets/icon--profile.png
